+++ game-core/src/state/mod.rs
@@ -1,0 +1,3 @@
+mod game;
+
+pub use self::game::Game;
