+++ game-core/src/system/mod.rs
@@ -1,6 +1,2 @@
 pub mod enemy;
-pub mod loader;
 pub mod player;
-
-pub use self::loader::load_texture;
-pub use self::loader::Loader;
